Improved isUnsigned to check attribute minimum more carefully
The previous implementation checked the validation predicate for a string containing literally ">= 0".  But if an attribute's minimum was, for example, 1 instead of 0, the test would fail because the string would be ">= 1".  But when a minimum is 1 we still want to use `unsigned`, but weren't getting `unsigned` because of this.

Updated the code to instead look for the string "SELF >=" and then inspect the number that comes after (if any). Then we can in fact correctly check that the minimus is >= 0, thus allowing `unsigned` to be used in those cases too.